bpf: Refactor object relationship tracking and fix dynptr UAF bug
Refactor object relationship tracking in the verifier by removing
dynptr_id and using ref_obj_id to track the parent object. By tracking
the parent object, it also fixes an use-after-free bug in BPF qdisc as
the verifier now properly invalidates skb dynptr and slices after the
underlying skb is freed through qdisc kfuncs.

To simplify object relationship tracking and allow the verifier to track
precise object relationship, change the semantics of id and ref_obj_id
of an object. Currently, "id" is assigned when an pointer type may be null
, "ref_obj_id" is used to track the underlying referenced object, and
"dynptr_id" is used for dynptr slices to track the parent dynptr. This
arrangement becomes insufficient to properly track object relationship.
Therefore, remove "dynptr_id" and used "ref_obj_id" universally to track
the parent object "id". For an obejct, if it is referenced, the "id" will
now be acquried through acquire_reference().

For release_reference(), this mean when invalidating an object, it needs
to also invalidate all dependent objects by traversing the subtree. This is
done using stack-based DFS to avoid recursive call chain of
release_reference() -> __unmark_stack_slots_dynptr() ->
release_reference().

CC: Kumar Kartikeya Dwivedi <memxor@gmail.com>
Fixes: 870c28588afa ("bpf: net_sched: Add basic bpf qdisc kfuncs")
Signed-off-by: Amery Hung <ameryhung@gmail.com>

---

This refactor is based on Kumar work in [0]

[0] https://lore.kernel.org/bpf/20250414161443.1146103-2-memxor@gmail.com/